Extract SSL configuration handling to a separate class (#123)
* Extract SSL configuration to a separate class

* KubernetesSsl -> Ssl, container -> local
